missingSummary improvements
Adds missingSummary to the package.

parseBranchingLogic is updated to be a little less aggressive on the lower casing.

constants.R is added to give access to field names created by REDCap. There's probably a better way to manage this than hard-coding constants, and I'm open to ideas.

also includes data and a data dictionary to test missingSummary on. Will develop tests separately.