Issue 32, checkbox_suffixes doesn't need a version argument
When looking at this I saw that this was a fourth way that I had deconstructed the choices, and was different than all the other three. I worked the issue in this branch since this one has the new fieldChoiceMapping function. This provides consistency across how we handle extracting the mapping.